Fixed black screens under Proton and various Windows configurations by using a separate overlay window.
Fixed video playback by adding 32-bit blitting support and removing the previous workaround.
Improved OpenGL renderer and software renderer performance.
Allow custom width/height settings for DirectDraw.
Added experimental ASLR/DEP support.
Added a safeguard against executing unimplemented functions.
Added minimal support for the Technical Demo.

+++ subtitans/iddrawpalette.h
@@ -23,8 +23,8 @@ namespace DDraw {
 
 		// Direct Draw Palette
 		virtual uint32_t __stdcall GetCaps(void*) = 0;
-		virtual uint32_t __stdcall GetEntries(uint32_t, uint32_t, uint32_t, uint8_t*) = 0;
+		virtual uint32_t __stdcall GetEntries(uint32_t, uint32_t, uint32_t, uint32_t*) = 0;
 		virtual uint32_t __stdcall Initialize(void*, uint32_t, void*) = 0;
-		virtual uint32_t __stdcall SetEntries(uint32_t, uint32_t, uint32_t, uint8_t*) = 0;
+		virtual uint32_t __stdcall SetEntries(uint32_t, uint32_t, uint32_t, uint32_t*) = 0;
 	};
 }
